1.1R1 Release Build
This build aims to make BLZ-OSPL more friendly to run code when it is
used as a library from another java program.

Full changelog:

+ Interpreter as a new static function called "executeCodeAsLibrary"
which takes an ArrayList of strings as the code to run as well as a List
of strings corresponding to the arguments and a new thing called a
BlzEventHandler

+ The BlzEventHandler, by default there is a StandAloneEventHandler that
will make standalone execution of code exactly the same as it was
before. But by using a different one, you can do things like intercept
output, pass input and handle exits.

* Everything that previously had a System.exit or a println now uses
this event handler to take care of that

* Changed some logic to handle not running from a file

* Trying to get any global variable information about the path will now
return "SOFTWARE" if it is running from a software environment

+ Added a library JAR

* Updated the runtime jar and executeExample bat