Implementation of the shunting yard algorithm
This is a massive rework of the underlying simple expression system that
increases performance, improves consistency and adds new features.

Instead of doing some bad string manipulation, I have implemented
Dijkstra's shunting yard algorithm. That uses a operation / operand
stack.

I have also added TONS of unit tests for both the Parser and Executor
for the new system.

The new system is not yet fully integerated, so lots of stuff is still
using the old SimpleExpression system. It should be phased out in the
near future.